requirements.txt
fastapi
uvicorn
python-multipart
pydantic
httpx
python-dotenv
google-genai
openai
request